refactor: Refactored to a single script for simplified UX
